pokedex_api/requirements.txt
typing_extensions==4.11.0
mysqlclient==2.2.4
tzdata==2024.1
uritemplate==4.1.1
sqlparse==0.5.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pytz==2024.1
inflection==0.5.1
PyJWT==2.8.0
Faker==24.11.0
six==1.16.0
packaging==24.0
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
mysqlclient==2.2.4